Changed Gamma Correction to spec by spec, adapted original scripts
- Changed Gammma Correction function to perform the algorithm spectra by spectra
- This change was necessary due to bug introduced in newer versions of mantid
- Adapted original scripts to use the same analysis inputs as the current scripts
- Current workflow is confusing, but consists of running current scripts through the
command line, runnning the standalone original scripts and finnaly running the standalone
comparison script.
