+++ EVSVesuvio/analysis_runner.py
@@ -11,7 +11,7 @@ from EVSVesuvio.scripts import handle_config
 
 
 def run():
-    scriptName =  handle_config.read_config_var('caching.experiment')
+    scriptName = handle_config.read_config_var('caching.experiment')
     experimentsPath = Path(handle_config.read_config_var('caching.location')) / "experiments" / scriptName # Path to the repository
     inputs_path = experimentsPath / "analysis_inputs.py"
     ai = import_from_path(inputs_path, "analysis_inputs")
